requirements.txt
torch
torchvision
deepspeed
toml
transformers==4.49.0
diffusers>=0.32.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
flash-attn
omegaconf
iopath
termcolor
hydra-core
easydict
ftfy
pytorch-optimizer
timm